requirements.txt
apache-flink
kafka-python==2.0.2
Flask==2.3.3
gunicorn==21.2.0
